Move default GATT callback logic to its own class
